include base interactor in the interactor module
