+++ lib/interactor.rb
@@ -2,6 +2,7 @@ require "interactor/context"
 require "interactor/error"
 require "interactor/hooks"
 require "interactor/organizer"
+require "interactor/base_interactor"
 
 # Public: Interactor methods. Because Interactor is a module, custom Interactor
 # classes should include Interactor rather than inherit from it.
@@ -21,6 +22,7 @@ module Interactor
     base.class_eval do
       extend ClassMethods
       include Hooks
+      include BaseInteractor
 
       # Public: Gets the Interactor::Context of the Interactor instance.
       attr_reader :context
